Consolidate redundant PATCH test and add HAS_PLONE_62 guard
