+++ src/plone/restapi/tests/test_content_patch.py
@@ -1,3 +1,4 @@
+from importlib import import_module
 from OFS.interfaces import IObjectWillBeAddedEvent
 from plone.app.testing import login
 from plone.app.testing import setRoles
@@ -18,6 +19,11 @@ import json
 import requests
 import transaction
 import unittest
+
+
+HAS_PLONE_62 = getattr(
+    import_module("Products.CMFPlone.factory"), "PLONE62MARKER", False
+)
 
 
 class TestContentPatch(unittest.TestCase):
@@ -197,3 +203,57 @@ class TestContentPatch(unittest.TestCase):
         self.assertEqual(204, response.status_code)
         transaction.begin()
         self.assertEqual("<p>example with '</p>", self.portal.doc1.text.raw)
+
+    @unittest.skipUnless(HAS_PLONE_62, "Requires Plone 6.2+")
+    def test_patch_image_redundant_no_event(self):
+        image_data = "R0lGODlhAQABAIAAAP///wAAACwAAAAAAQABAAACAkQBADs="
+        response = requests.post(
+            self.portal.absolute_url(),
+            headers={"Accept": "application/json"},
+            auth=(SITE_OWNER_NAME, SITE_OWNER_PASSWORD),
+            json={
+                "@type": "Image",
+                "image": {
+                    "data": image_data,
+                    "encoding": "base64",
+                    "content-type": "image/gif",
+                    "filename": "test.gif",
+                },
+            },
+        )
+        self.assertEqual(201, response.status_code)
+        image_url = response.json()["@id"]
+        transaction.commit()
+
+        # Track ObjectModifiedEvent
+        sm = getGlobalSiteManager()
+        fired_events = []
+
+        def record_event(event):
+            fired_events.append(event)
+
+        sm.registerHandler(record_event, (IObjectModifiedEvent,))
+
+        # Patch with same data
+        response = requests.patch(
+            image_url,
+            headers={"Accept": "application/json"},
+            auth=(SITE_OWNER_NAME, SITE_OWNER_PASSWORD),
+            json={
+                "image": {
+                    "data": image_data,
+                    "encoding": "base64",
+                    "content-type": "image/gif",
+                    "filename": "test.gif",
+                },
+            },
+        )
+        self.assertEqual(204, response.status_code)
+
+        # In current Plone, this will be 1 because of identity mismatch
+        # We want it to be 0
+        self.assertEqual(
+            len(fired_events), 0, "ObjectModifiedEvent was fired for redundant PATCH"
+        )
+
+        sm.unregisterHandler(record_event, (IObjectModifiedEvent,))
